feat: 구입 금액에 맞게 랜덤하게 로또 생성하는 기능 구현 완료
- docs: '로또 구입 금액으로 랜덤하게 로또 생성하기' 항목 완료
- test: 정상 상황에서 금액에 따른 로또 생성 테스트 완료
- docs: '구입 금액을 1000으로 나누어 생성할 로또 개수 구하기' 항목 완료
- test: 정상 구매금액의 경우 로또개수 정상 테스트 완료

+++ src/main/java/lotto/view/InputView.java
@@ -5,11 +5,11 @@ import camp.nextstep.edu.missionutils.Console;
 public class InputView {
     private static final String PURCHASE_AMOUNT_MESSAGE = "구입금액을 입력해 주세요.";
 
-    public int readPurchaseAmount() {
+    public long readPurchaseAmount() {
         System.out.println(PURCHASE_AMOUNT_MESSAGE);
         String userPurchaseAmount = Console.readLine();
         checkUserInputIsNull(userPurchaseAmount);
-        
+
         return checkPurchaseAmountIsNum(userPurchaseAmount);
     }
 
@@ -19,9 +19,9 @@ public class InputView {
         }
     }
 
-    private int checkPurchaseAmountIsNum(String purchaseAmount) {
+    private long checkPurchaseAmountIsNum(String purchaseAmount) {
         try {
-            return Integer.parseInt(purchaseAmount);
+            return Long.parseLong(purchaseAmount);
         } catch (NumberFormatException e) {
             throw new IllegalArgumentException("[ERROR] 숫자를 입력해주세요.");
         }
